Split socket & message functions into new socket.reducer
