+++ src/malcolm/reducer/malcolmReducer.js
@@ -1,26 +1,19 @@
 import {
-  MalcolmSend,
-  MalcolmError,
   MalcolmAttributeFlag,
   MalcolmNavigationPathUpdate,
-  MalcolmDisconnected,
-  MalcolmReturn,
   MalcolmUpdateBlockPosition,
   MalcolmShiftButton,
-  MalcolmSocketConnect,
   MalcolmSelectPortType,
 } from '../malcolm.types';
 import blockUtils from '../blockUtils';
 import { AlarmStates } from '../../malcolmWidgets/attributeDetails/attributeAlarm/attributeAlarm.component';
 import NavigationReducer from './navigation.reducer';
-import AttributeReducer, {
-  updateAttribute,
-  pushToArchive,
-} from './attribute.reducer';
+import AttributeReducer from './attribute.reducer';
 import layoutReducer, { LayoutReduxReducer } from './layout/layout.reducer';
 import methodReducer from './method.reducer';
 import tableReducer from './table.reducer';
 import BlockReducer from './block.reducer';
+import SocketReducer from './socket.reducer';
 
 export const ARCHIVE_BUFFER_LENGTH = 1000; // length of circular buffer used for archiving
 export const ARCHIVE_REFRESH_INTERVAL = 2.0; // minimum time in seconds between updates of displayed archive data
@@ -59,35 +52,6 @@ const initialMalcolmState = {
     inDeleteZone: false,
   },
 };
-
-function updateMessagesInFlight(state, action) {
-  const newState = state;
-
-  if (
-    action.payload.typeid !== 'malcolm:core/Subscribe:1.0' ||
-    !Object.keys(state.messagesInFlight).some(
-      m =>
-        state.messagesInFlight[m] !== undefined &&
-        state.messagesInFlight[m].path.join() === action.payload.path.join()
-    )
-  ) {
-    newState.messagesInFlight = {
-      ...state.messagesInFlight,
-    };
-    newState.messagesInFlight[action.payload.id] = action.payload;
-  }
-
-  return newState;
-}
-
-function stopTrackingMessage(state, action) {
-  const filteredMessages = { ...state.messagesInFlight };
-  delete filteredMessages[action.payload.id];
-  return {
-    ...state,
-    messagesInFlight: filteredMessages,
-  };
-}
 
 function setFlag(state, path, flagType, flagState) {
   if (path.length === 1) {
@@ -153,182 +117,6 @@ function setFlag(state, path, flagType, flagState) {
   return updatedState;
 }
 
-function setDisconnected(state) {
-  const blocks = { ...state.blocks };
-  Object.keys(blocks).forEach(blockName => {
-    if (Object.prototype.hasOwnProperty.call(blocks[blockName], 'attributes')) {
-      const attributes = [...state.blocks[blockName].attributes];
-      for (let attr = 0; attr < attributes.length; attr += 1) {
-        if (Object.prototype.hasOwnProperty.call(attributes[attr], 'raw')) {
-          if (
-            Object.prototype.hasOwnProperty.call(attributes[attr].raw, 'meta')
-          ) {
-            attributes[attr].raw = {
-              ...attributes[attr].raw,
-              meta: {
-                ...attributes[attr].raw.meta,
-                writeable: false,
-              },
-            };
-          }
-          if (
-            Object.prototype.hasOwnProperty.call(attributes[attr].raw, 'alarm')
-          ) {
-            attributes[attr].raw = {
-              ...attributes[attr].raw,
-              alarm: {
-                ...attributes[attr].raw.alarm,
-                severity: AlarmStates.UNDEFINED_ALARM,
-                message: 'Websocket connection to Malcolm lost!',
-              },
-            };
-            if (
-              state.blockArchive[blockName] &&
-              state.blockArchive[blockName].attributes[attr] &&
-              state.blockArchive[blockName].attributes[attr].alarmState.get(
-                state.blockArchive[blockName].attributes[
-                  attr
-                ].alarmState.size() - 1
-              ) !== AlarmStates.UNDEFINED_ALARM
-            ) {
-              const { timeStamp } = attributes[attr].raw;
-              pushToArchive(
-                state.blockArchive[blockName].attributes[attr],
-                {
-                  raw: {
-                    timeStamp,
-                    value: attributes[attr].raw.value,
-                  },
-                },
-                AlarmStates.UNDEFINED_ALARM
-              );
-            }
-          }
-        }
-      }
-      blocks[blockName] = { ...state.blocks[blockName], attributes };
-    }
-  });
-  return {
-    ...state,
-    blocks,
-    counter: 0,
-  };
-}
-
-export const setErrorState = (state, id, errorState, errorMessage) => {
-  const matchingMessage = state.messagesInFlight[id];
-  const path = matchingMessage ? matchingMessage.path : undefined;
-  if (path) {
-    const blockName = path[0];
-    const attributeName = path[1];
-
-    const matchingAttributeIndex = blockUtils.findAttributeIndex(
-      state.blocks,
-      blockName,
-      attributeName
-    );
-    const blocks = { ...state.blocks };
-    if (matchingAttributeIndex >= 0) {
-      const { attributes } = state.blocks[blockName];
-      attributes[matchingAttributeIndex] = {
-        ...attributes[matchingAttributeIndex],
-        calculated: {
-          ...attributes[matchingAttributeIndex].calculated,
-          errorState,
-          errorMessage,
-          dirty: errorState,
-          forceUpdate: !errorState,
-          alarms: {
-            ...attributes[matchingAttributeIndex].calculated.alarms,
-            dirty: errorState ? AlarmStates.DIRTY : null,
-            errorState: errorState ? AlarmStates.MAJOR_ALARM : null,
-          },
-        },
-      };
-      blocks[blockName] = { ...state.blocks[blockName], attributes };
-    }
-    return {
-      ...state,
-      blocks,
-    };
-  }
-  return state;
-};
-
-function handleReturnMessage(state, action) {
-  const newState = setErrorState(state, action.payload.id, false, 'Successful');
-  return stopTrackingMessage(newState, action);
-}
-
-const handleErrorMessage = (state, action) => {
-  const matchingMessage = state.messagesInFlight[action.payload.id];
-  let updatedState = { ...state };
-  if (matchingMessage && matchingMessage.path) {
-    const blockName = matchingMessage.path[0];
-    const matchingAttributeIndex = blockUtils.findAttributeIndex(
-      state.blocks,
-      blockName,
-      matchingMessage.path[1]
-    );
-    const attribute =
-      matchingAttributeIndex > -1
-        ? state.blocks[blockName].attributes[matchingAttributeIndex]
-        : undefined;
-    if (
-      attribute &&
-      attribute.raw &&
-      attribute.raw.meta &&
-      attribute.raw.meta.tags &&
-      attribute.raw.meta.tags.some(t => t === 'widget:flowgraph')
-    ) {
-      // reset the layout
-      const id = attribute.id === undefined ? attribute.id : action.payload.id;
-      updatedState = updateAttribute(state, {
-        id,
-        delta: true,
-      });
-    } else if (
-      attribute &&
-      attribute.raw.typeid === 'malcolm:core/Method:1.0'
-    ) {
-      const attributes = [...state.blockArchive[blockName].attributes];
-      const archive = { ...attributes[matchingAttributeIndex] };
-      const runParams = archive.value.pop();
-      const localRunTime = archive.timeStamp.pop();
-      archive.value.push({
-        ...runParams,
-        returned: { error: action.payload.message },
-        returnStatus: `Failed: ${action.payload.message}`,
-      });
-      archive.timeStamp.push({ ...localRunTime, localReturnTime: new Date() });
-      archive.alarmState.push(AlarmStates.MAJOR_ALARM);
-      attributes[matchingAttributeIndex] = archive;
-      updatedState.blockArchive[blockName] = {
-        ...state.blockArchive[blockName],
-        attributes,
-      };
-    }
-  }
-
-  updatedState = setErrorState(
-    updatedState,
-    action.payload.id,
-    true,
-    action.payload.message
-  );
-  return stopTrackingMessage(updatedState, action);
-};
-
-const updateSocket = (state, payload) => {
-  const { worker } = payload;
-  worker.postMessage(`connect::${payload.socketUrl}`);
-
-  return {
-    ...state,
-  };
-};
-
 const updateLayoutOnState = state => {
   const updatedState = state;
   const layoutUpdates = layoutReducer.updateLayoutAndEngine(updatedState);
@@ -344,6 +132,7 @@ const malcolmReducer = (state = initialMalcolmState, action = {}) => {
   updatedState = tableReducer(updatedState, action);
   updatedState = LayoutReduxReducer(updatedState, action);
   updatedState = BlockReducer(updatedState, action);
+  updatedState = SocketReducer(updatedState, action);
 
   switch (action.type) {
     case MalcolmAttributeFlag:
@@ -353,24 +142,6 @@ const malcolmReducer = (state = initialMalcolmState, action = {}) => {
         action.payload.flagType,
         action.payload.flagState
       );
-
-    // <message reducer>
-    case MalcolmSend:
-      return updateMessagesInFlight(updatedState, action);
-
-    case MalcolmError:
-      return handleErrorMessage(updatedState, action);
-
-    case MalcolmReturn:
-      return handleReturnMessage(updatedState, action);
-
-    case MalcolmSocketConnect:
-      return updateSocket(updatedState, action.payload);
-
-    case MalcolmDisconnected:
-      return setDisconnected(updatedState);
-
-    // </message reducer>
 
     case MalcolmNavigationPathUpdate:
       updatedState = NavigationReducer.updateNavigationPath(
